// src/types/pph-4-ayat-2-type.ts
export type TPPh4Ayat2FormSchema = {
  jenis_penghasilan: string;
  uraian_kegiatan: string;
  no_pengajuan: string;
  target_penerima: string;
  nama_penerima: string;
  nama_bank: string;
  no_rekening: string;
  nama_rekening: string;
  npwp: string;
  kode_objek_pajak: string;
  penghasilan_bruto: number;
  tarif_pajak: number;
  potongan_pajak: number;
  penghasilan_diterima: number;
  pic_pencairan_penghasilan: string;
  invoice: string;
  faktur_pajak: string | null;
  dokumen_kerjasama_kegiatan: string | null;
  idl: string;
};

export type TPPh4Ayat2Model = {
  kode_kegiatan_penghasilan_badan_usaha: string;
  jenis_penghasilan: string;
  uraian_kegiatan: string;
  no_pengajuan: string;
  target_penerima: string;
  nama_penerima: string;
  nama_bank: string;
  no_rekening?: string | null;
  nama_rekening?: string | null;
  npwp?: string | null;
  nama_narahubung: string;
  kode_objek_pajak: string;
  penghasilan_bruto: number;
  tarif_pajak: number;
  potongan_pajak: number;
  penghasilan_diterima: number;
  pic_pencairan_penghasilan: string;
  invoice: string;
  faktur_pajak: string;
  dokumen_kerjasama_kegiatan: string;
  idl: string;
  kode_jenis_pajak: number;
  status: string;
  tanggal_input: Date;
};